Use public url instead of s3 bucket (#154)
Use the PUBLIC_URL env var to set URLs for the various bundles, shims, replacing REACT_APP_S3_BUCKET to allow assets to be served from sub directories in our S3 bucket.

* Moves web component port to 3001 from 9000 to avoid clashes with auth
* Replace REACT_APP_S3_BUCKET with PUBLIC_URL everywhere
* Rename the web-component bundle.js to web-component.js
* Use the Webpack HTML plugin to build the web component index.html as web-component.html, and inject the bundle correctly into the body.
* The web-component index is now build/web-component.html.
* Adds Cypress tests to GH actions
* Add a delay to CircleCI to start cypress only after the services had started.

The web component works, served straight from S3, picking up the asset paths correctly:  However, the editor itself does not, because it makes the assumption it is being served form the root directory. I think that is too big to cover off in this PR, and maybe we should address it in a different way.