Use nginx instead of webrick when hosting on Heroku (#156)
This uses the heroku NGINX buildpack to serve up the static files instead of a ruby process.

This is better in prod than webrick, as it is a proper webserver that can handle more than one request at a time. This will do while we sort out Cloudflare buckets etc.