Sidebar fixes & updates (#720)
## What does this PR include

Added:
- SVGO config file

Updated:
- Stopped SVGO from removing svg viewBoxes
- Sidebar - Refactored styles to be closer to designs
- Sidebar - Refactored theming to fix inconsistencies between themes
- Sidebar - Added hover styles to collapse and close buttons
- Sidebar - Option/Tab style

Fixed:
- Edit icon not showing in ContextMenu (Fixes
https://github.com/RaspberryPiFoundation/editor-ui/issues/721)
- Updated sidebar and file icons to correct size
- Fixed stretched icon tabs
- FileMenu alignment

## Futher details

> - SVGO config file

SVGO was removing viewBox from SVG icons, which means icons were not
sized correctly and would not scale when you what to change the default
size. I added a config file for SVGO and included that in the webpack
config.

> - Sidebar - Refactored theming to fix inconsistencies between themes

I've tried a different way of defining the themes as the current way
seemed to create a lot of duplication and was leading to layout changes
between the themes.

With how I've got it now, themed elements are defined once and colours
are updated with the custom properties at the top. I think it makes it
easier to read and maintain too? On another note I think this route
would work well with design tokens too.

> - Sidebar - Added hover styles to collapse and close buttons

I used the same wrapper class used on other sidebar elements - I think
this is all okay?

## Still to do

- [x] I need some help with linting
- [x] Would any tests need updating?
- [x] Check no other icons are broken, I think they all look good.
- [x] Check if there are any redudant classes on sidebar element
- [x] Double check image icon in sidebar looks good on both themes, I
wasn't sure how to demo this locally.

Some things left to do, but that could be in a different PR:
- [ ] Update spacing variables to be using design system
- [ ] Counter balance the extra width that is added to the sidebar-bar
by all the borders
- [ ] Close button size is slight different that the others.
- [ ] Remove the line between the selected tab and the panel so they
look connected - Pseudo-Elements for this?

---------

Co-authored-by: MagdalenaJadach <magdalena@raspberrypi.org>