Add output_only property to web component
Based on changes in this commit [1].

This changes `WebComponentProject` so it displays only the `Output`
component instead of the `Project` or `MobileProject` component. This
closely matches the existing behaviour of the `EmbeddedViewer`
component. And in fact we also style the `Output` component like the
latter by wrapping it in the "embedded-viewer" CSS class and importing
the `EmbeddedViewer.scss` styles.

The immediate motivation for this change is to support the Block-to-Text
app [2], but we also hope to replace the `EmbeddedViewer` route with the
web component in some kind of embedded mode. The exercise in the
Block-to-Text app uses this attribute [3] to display the output
generated from running the code.

In order to use the `Output` component in this new web component
context, we've added `embedded` & `browserPreview` properties to it. I
think this is because we need to be able to set them explicitly in the
context of the web component since we can't (or don't want to) rely on
setting them from the query string as compared with using the
`EmbeddedViewer` in the context of an `iframe`.

However, it's slightly odd to me that we've done this for
`browserPreview` at this point, because we explicitly set it to `false`
in `WebComponentProject` which is the default. But perhaps this is
thinking ahead a bit to wanting to make these properties available as
web component attributes so that we can do away with using
`EmbeddedViewer` in an `iframe`.

I've also added some unit tests - I've added `data-testid` attributes to
a few elements to achieve this. I plan to add an e2e spec in a
subsequent commit which will test all the new attributes on the web
component.

[1]: https://github.com/RaspberryPiFoundation/editor-ui/pull/782/commits/28de5b8d0052ac47499fe808aaa50d8a80f335fd
[2]: https://github.com/RaspberryPiFoundation/block-to-text-alpha/
[3]: https://github.com/RaspberryPiFoundation/block-to-text-alpha/blob/d16a8ecc1d3da202363fb16db31f4bf93f46aa04/src/components/Editor.js#L18