Extracting user from localStorage (#852)
closes #743

The idea is to intercept editor actions (i.e. actions that could call
the API) and update the user object in state prior to that call being
made. That way if the user object changes for whatever reason (like cred
renewal) the latest user object will be used for the API call