Add runCode, stopCode & rerunCode to web component
Based on [1] & #899.

The motivation for this change is to support the Block-to-Text app [2]
which uses the `rerunCode` method in its exercise to allow the user to
run the code from the button outside the web component [3].

This mainly adds the `runCode`, `stopCode`, & `rerunCode` methods to the
web component, but it also adds a check on the `webComponent` boolean
state in `Output` to determine whether or not to render the `RunBar`. I
don't really understand why this is needed, but having chatted to
@loiswells97 we've got a nagging feeling there's a use case that does
need it, so I'm going to leave it in plae for now. I've added a unit
test to illustrate the new behaviour.

I plan to add an e2e spec in a subsequent commit which will test the new
methods on the web component.

[1]: https://github.com/RaspberryPiFoundation/editor-ui/pull/782/commits/f9086a88bb343ef7ff4ce8623a64d1162191c293
[2]: https://github.com/RaspberryPiFoundation/block-to-text-alpha/
[3]: https://github.com/RaspberryPiFoundation/block-to-text-alpha/blob/d16a8ecc1d3da202363fb16db31f4bf93f46aa04/src/components/Step.js#L87