Update webpack-dev-server from 3.11.1 to 4.0.0
We need to be able to conditionally set headers for pyodideWorker.js in
pull request https://github.com/RaspberryPiFoundation/editor-ui/pull/893
but this feature isn’t supported in webpack-dev-server 3.x.x. We
therefore need to upgrade to version 4.0.0 which allows a function to be
provided in the headers option.

I’ve tried to follow the migration guide here:
https://github.com/webpack/webpack-dev-server/blob/master/migration-v4.md