Update yarn from version 1.22.21 to 3.4.1 and webpack-dev-server from 3.11.1 to 4.0.0 (#895)
We need to be able to conditionally set headers for pyodideWorker.js in
pull request https://github.com/RaspberryPiFoundation/editor-ui/pull/893
but this feature isn’t supported in webpack-dev-server 3.x.x. We
therefore need to upgrade to version 4.0.0 which allows a function to be
provided in the headers option. When attempting this upgrade, I ran into
[this issue](https://github.com/storybookjs/storybook/issues/22431) and
only managed to resolve it by also updating yarn to version 3.

I’ve tried to follow the migration guide here:

https://github.com/webpack/webpack-dev-server/blob/master/migration-v4.md