Write components to the Wasm in-memory filesystem before running programs
https://pyodide.org/en/stable/usage/api/js-api.html#pyodide.FS

In the future, we could open this up to support other file types
(not just .py, .txt and .csv) and we could allow users to create files
with Python and then save the result to a component in the project, e.g.
to display an image of the Mandelbrot set.