Implement visualOutput.handleVisual to display graphical output
I originally tried implementing this with redux by dispatching an event
containing the ‘content’ object but this isn’t always serialisable, e.g.
for turtle it is a JavaScript Map object. Redux prints a warning if the
object isn’t serialisable. Therefore, instead, use a channel pattern
that allows the sub-component (VisualOutput) to subscribe to the
handleVisual method calls and update its DOM.