Add support for stdin control keys
Previously, if you ran the following program in the editor using Pyodide
there was no way to cleanly exit:

```python
import sys

for line in sys.stdin:
    print(line)
```

This wasn’t an issue before because Skulpt only supported the input()
function which scans to the end of the line and does not look for EOF
when the standard input is closed.

To support this, add the ability to press control-d which sets the first
element in the stdinBuffer to -1. This signals to the worker that stdin
is closed, which it signals to Pyodide by returning 0 from read(buffer).