Dynamically show the VisualOutputPane when a visual program is run
Previously, we had to manually add ?show_visual_output=true to the URL
to show the pane. We now set showVisual to true when the first
handleVisual message is received and use a ‘showing’ state variable on
the visual objects to determine whether we have already rendered the
visual to avoid re-renders.