Load PyodideWorker.mock.js in the Jest tests
Jest doesn’t support Web Workers and was erroring because the
importScripts function isn’t available. Best practice seems to be to
mock the module and returned canned responses. Currently, we don’t have
Jest tests for the experimental PyodideRunner.js so an empty
implementation is sufficient but we could change this later.

https://stackoverflow.com/questions/42567535#answer-62436219