Send error details in "editor-runCompleted" event
Based on #915. This stores a structured version of the error in
`errorDetails` in the Redux store in the `PythonRunner` `handleError`
function and sends it in the payload of the "editor-runCompleted" custom
event from the `WebComponentProject` component.

The motivation for this change is to support the Block-to-Text app [1]
which uses this to display an error message to the user in its
exercise [2].

It also hides the `ErrorMessage` component in `PythonRunner` when
`isSplitView` is `false`.

I've added some unit tests for the new behaviour. I plan to add an e2e
spec in a subsequent commit which will test all the new attributes on
the web component.

[1]: https://github.com/RaspberryPiFoundation/block-to-text-alpha
[2]: https://github.com/RaspberryPiFoundation/block-to-text-alpha/blob/d16a8ecc1d3da202363fb16db31f4bf93f46aa04/src/components/Step.js#L53