Specify code output panels in web component
Based on #909. The new `output_panels` attribute on the web component
allows specific panels to be selected when the `output-only` web
component attribute is set to `true`. Currently there are two panels:
the "text" panel and the "visual" panel. You can select either or both.
Note that this functionality has only currently been implemented for the
`PythonRunner` and not the `PyodideRunner` or the `HtmlRunner`.

The motivation for this change is to support the Block-to-Text app [1].
The exercise in the alpha version of this app displays just the "visual"
panel [2].

I've added a couple of unit tests in `PythonRunner.test.js` to test the
new behaviour. And I plan to add an e2e spec in a subsequent commit
which will test all the new attributes on the web component.

[1]: https://github.com/RaspberryPiFoundation/block-to-text-alpha/
[2]: https://github.com/RaspberryPiFoundation/block-to-text-alpha/blob/d16a8ecc1d3da202363fb16db31f4bf93f46aa04/src/components/Editor.js#L19