Update references to the application port
We think these are vestiges of the `editor-ui` setup where the app runs
on port 3000 locally and 3010 in Docker. The `editor-standalone` app
runs on port 3012 both locally & in Docker.

The `yarn:start` script used in the "Cypress run" step of the
`ci-cd.yml` workflow starts the app on port 3012, so we need to change
the `wait-on` option so it waits for the app to come up on that port.

The `PUBLIC_URL` only seems to be used to in the URL for the
`manifest.json` URL, but it seems to make sense to make it consistent
here.

The `baseUrl` set in `cypress/e2e/spec.cy.js` &
`cypress/e2e/spec-html.cy.js` needs to be changed to use port 3012 so
that the Cypress tests access the app on the correct port. There might
be a better way to configure this more generically rather than in each
test, but we've gone for the smallest change that fixes the problem in
this commit.

Co-authored-by: Chris Roos <chris.roos@gofreerange.com>
Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>