Look inside the shadow DOM in the Cypress tests
The "renders the web component" test was failing because we include the
editor using the "editor-wc" web component which creates elements in a
shadow DOM and these elements are not accessible via the Cypress' `get`
command. We need to use Cypress' `shadow` command [1] to look inside the
shadow DOM. The test should now pass.

[1] https://docs.cypress.io/api/commands/shadow

Co-authored-by: Chris Roos <chris.roos@gofreerange.com>
Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>