Source web component from staging in CI workflow
Following the pattern we've used for the API endpoint (i.e.
`REACT_APP_API_ENDPOINT`), this points the app at the static staging
instance of the `editor-ui` app where it can obtain the editor web
component. This avoids having to run the `editor-ui` app as part of the
CI workflow.

Co-authored-by: Chris Roos <chris.roos@gofreerange.com>
Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>