Remove all but one of the Cypress tests
All of these tests were testing the behaviour of the `editor-ui` web
component rather than that of the `editor-standalone` app. We plan to
retain the tests in the `editor-ui` repo. We've just kept a single test
in `editor-standalone` which checks that the editor web component has
loaded/rendered OK. Note that the latter isn't working yet, but we'll
fix that in a subsequent commit.

Co-authored-by: Chris Roos <chris.roos@gofreerange.com>
Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>