Avoid react-modal warnings in BetaModal test
We were seeing the following warnings:

    Warning: react-modal: App element is not defined. Please use
    `Modal.setAppElement(el)` or set `appElement={el}`. This is needed so
    screen readers don't see main content when modal is opened. It is not
    recommended, but you can opt-out by setting `ariaHideApp={false}`.

These are caused by our setting of `appElement` to either `<editor-wc>`
or `#app` in `GeneralModal`. In our test we render the modal component
in a div with id of `app` but that isn't available in the DOM at the
point React attempts to render the component. By adding a div with id of
`app` to the DOM before all the tests we can avoid this warning.

Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>