Avoid react-modal warnings in RenameFileModal test
We were seeing the following warnings:

    Warning: react-modal: App element is not defined. Please use
    `Modal.setAppElement(el)` or set `appElement={el}`. This is needed so
    screen readers don't see main content when modal is opened. It is not
    recommended, but you can opt-out by setting `ariaHideApp={false}`.

These are caused by our setting of `appElement` to either `<editor-wc>`
or `#app` in `GeneralModal`. In our test we render the modal component
in a div with id of `app` but that isn't available in the DOM at the
point React attempts to render the component. By adding a div with id of
`app` to the DOM before all the tests we can avoid this warning.

Note that we're now having to pass `{ hidden: true }` into the call to
`screen.getByRole` to find the dialog. This is because react-modal adds
an `aria-hidden=true` attribute to the `appElement`, which is an element
with id of `app` in our case. We think this might be highlighting a
potential accessibility problem in our use of React Modal because we're
nesting the modal within that same `appElement`, which the docs[1]
advise against:

> If you do this, please ensure that your app element is set correctly.
> The app element should not be a parent of the modal, to prevent modal
> content from being hidden to screenreaders while it is open.

[1]: https://reactcommunity.org/react-modal/#custom-parent

Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>