Import SVGs correctly in ProjectBar
To fix the following class of warnings we're seeing when running the
tests for this component:

    Warning: The tag <download.svg> is unrecognized in this browser. If
    you meant to render a React component, start its name with an uppercase
    letter.

Note that we've also changed the expected type of `Icon` that's being
passed to `DownloadButton`. It was previously requiring a function but
we were passing it a string, which was showing warnings in the tests.
It's now requiring an object and we're passing it a React Component so
we're no longer seeing warnings when running tests for this component.

Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>