Show Apollo Client error messages in tests
I find it more useful to see the error message in the output, than a URL
encoded version of an error message that uses
https://www.apollographql.com/docs/react/errors to display it!

The docs in the link above suggest that it should be enough to use
`loadErrorMessages()` on its own but that doesn't appear to do anything!
Using both `loadDevMessages()` and `loadErrorMessages()` does show us
the error messages in the test output so we're going to stick with using
both.