Fail fast if there are console warnings/errors in tests
By default Jest tests don't fail if there are console warnings/errors.
We've ended up with quite a few warnings/errors in our tests, at least
one of which was causing a problem in production[1]. We think it's
better to treat these warnings/errors as things that need to be
addressed and so are using this library to fail the tests. It will fail
on `console.error` and `console.warn` by default.

[1]: https://github.com/RaspberryPiFoundation/editor-standalone/pull/49

Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>