Add handler for editor-logIn event
When a user clicks the "Log in to save" button inside the editor web
component a custom `editor-logIn` event is fired. Before this commit
we were not listening for that event so clicking the button did
nothing.

We've added an event handler inspired by the one in
`projects-ui`[1]. Our handler calls the existing `login` function
that is also used in this site's drop down menu. It seems to work
correctly and redirect the user back to the correct project when we
pass the `project` and `location` arguments to `login` so for now
we've ignored the other arguments that this function can take.

[1] https://github.com/RaspberryPiFoundation/projects-ui/blob/28622ba1732249e516623e1ddd34ea3fde2c6a08/src/containers/ProjectEditorContainer/ProjectEditorContainer.js#L200

Co-authored-by: Chris Roos <chris.roos@gofreerange.com>