Redirect to new identifier URL after project remix
See this issue [1].

The `loading` and `project` parts of the editor state don't seem to be
available outside the editor web component and so unlike in the
integrated version of the editor in `editor-ui` the original `useEffect`
hook wasn't being triggered here in `editor-standalone` where we're
using the web component.

I've fixed this by making a new custom event
("editor-projectIdentifierChanged") available from the web component and
adding an event listener in `ProjectComponentLoader` which redirects to
the remixed project identifier URL.

I notice there is a TODO above some nearby/related commented-out code
which says: "sort out loading and error states", but I'm not exactly
sure what that means.

[1]: https://github.com/RaspberryPiFoundation/editor-standalone/issues/55