Link from global nav menu to user's school dashboard
This makes a request to editor-api for the schools that are accessible
to the logged in user. The API allows for multiple schools to be
associated with a single user but our designs currently assume a single
school per user, which is why we're arbitrarily taking the first one.

I've extracted `renderLoginMenu` in LoginMenu.test.jsx so that I can
stub the response from `getMySchool` before rendering in the "links to
the user's school in the menu" test. An alternative would've been to
configure the stub for all tests but I prefer the minimal test setup
required for a particular test/assertion.

According to the logic in our Ability class in editor-api[1], schools
are accessible to a user if the user is associated with the school and
has any of the school-owner, school-teacher or school-student roles.

You can demo this in development by:

1. Add the school-owner role to the john-doe user in dev by executing
   the following command in the profile app:

        node profile-cli assign-role john.doe@example.com school-owner editor-dev -c

2. Add the dummy school into the editor-api by executing the following
   in a Rails console:

        School.create!(
          name: "School name",
          website: "website",
          address_line_1: "address-line-1",
          municipality: "municipality",
          country_code: "GB",
          id: "12345678-1234-1234-1234-123456789abc"
        )

    Note that the ID in the school above has to match that in
`User#temporarily_add_organisations_until_the_profile_app_is_updated`[2]

[1]: https://github.com/RaspberryPiFoundation/editor-api/blob/0f63ea74fbe13e9878909b88368490a3988f03d3/app/models/ability.rb#L37
[2]: https://github.com/RaspberryPiFoundation/editor-api/blob/0f63ea74fbe13e9878909b88368490a3988f03d3/app/models/user.rb#L120

Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>