Dispatch event when project identifier changes
See this issue [1].

The integrated version of the editor in `editor-ui` redirects to the new
project identifier URL after a project is remixed. However, the
`editor-standalone` version which uses the editor web component was
unable to detect changes in the project identifier and so the redirect
was not working.

I've addressed this by making a new custom event
("editor-projectIdentifierChanged") available from the web component
with the idea that we'll add an event listener in
`ProjectComponentLoader` which redirects to the remixed project
identifier URL. The custom event is not dispatched when the project
identifier is not set, because that's not currently useful in
`editor-standalone`, but it would be easy enough to change.

This change should not change the behaviour of the `projects-ui` app
which also uses the web component, because it is not listening for the
new custom event. However, it would be easy enough to change that if it
was desirable.

Note that I've added a separate integration test for
`WebComponentProject`, because I wanted to use a real Redux store rather
than the mock store that's being used in the unit test. This is the
approach recommended in the Redux docs [2]:

> Our recommendation for testing Redux-connected React components is via
> integration tests that include everything working together, with
> assertions aimed at verifying that the app behaves the way you expect
> when the user interacts with it in a given manner.

[1]: https://github.com/RaspberryPiFoundation/editor-standalone/issues/55
[2]: https://redux.js.org/usage/writing-tests#integration-testing-connected-components-and-redux-logic