Remove broken `format` script
This script is broken, because no `.prettierrc` file is included in the
repo.

We're relying on the `prettier` plugin for `eslint` to do code
formatting & checking, e.g. in `.github/workflows/ci-cd.yml` where we
invoke `yarn lint` [1] to check the formatting.

Similarly, the `lint:fix` script can be used to apply formatting. So the
`format` script is redundant and potentially confusing to new
developers.

[1]: https://github.com/RaspberryPiFoundation/editor-standalone/blob/2621a6a28acba956977882e66d13eeb23cc716f2/.github/workflows/ci-cd.yml#L32-L33