Add /schools/<id> route and basic School component
To form the basis of the school dashboard page.

This is _very_ basic but hopefully provides enough of a skeleton for
other people to build upon.

We're currently handling some API errors (403 and 404) in the School
component but I wonder whether it'd be better to handle these in
somewhere like apiCallHandler. Hopefully our continued use of the API
will help inform such a decision.

GitHub issue: #43