Remove broken `format` script (#991)
This script is broken, because no `.prettierrc` file is included in the
repo.

We're relying on the `prettier` plugin for `eslint` to do code
formatting & checking, e.g. in `.github/workflows/ci-cd.yml` where we
invoke `yarn lint` [1] to check the formatting.

Similarly, the `lint:fix` script can be used to apply formatting. So the
`format` script is redundant and potentially confusing to new
developers.

[1]:
https://github.com/RaspberryPiFoundation/editor-ui/blob/5197facfc6253be548205d24e89478af286eb384/.github/workflows/ci-cd.yml#L31-L32