Add load_remix_disabled attribute to web component
This work came out of this issue [1].

This means that it is now possible to configure the web component not to
load a logged-in user's remixed version of the project by setting
`load_remix_disabled` to `true`.

We want to use this in editor-standalone with the web component editor
so that we reproduce the behaviour currently in editor-ui with the
integrated editor.

When the new attribute is set to `true`, the `loadRemix` property passed
to `useProject` in `WebComponentLoader` is forced to be `false` so that
the remixed version of the project will not be loaded.

The name of the new attribute isn't ideal - I would rather have used
`load_remix_enabled` - however, boolean attributes of the web component
default to `false` and I want the default behaviour to remain unchanged
to avoid breaking the projects app.

[1]: https://github.com/RaspberryPiFoundation/editor-standalone/issues/55.