Invalidate cached project when project is remixed
This fixes an issue which you can see if you follow these steps:

1. Login
2. Visit a project you do not own
3. Edit the code
4. Save the project (thus remixing it)
5. Visit the original project URL again
6. See your changes in the code when you should see the original version

This was happening because after step 3 the auto-save functionality
stored the project (with your code changes) in local storage against the
original project identifier. When the project was saved, the remixed
version was created in the database, but the cached version was left in
local storage. When you re-visit the original project URL, the project
cached in local storage was loaded, because the URL path matches the
local storage key, i.e. the original project identifier.

The fix is to invalidate the cached project after it is successfully
remixed in a similar way to how the cached project is invalidated after
it is successfully saved [1]. In this case, since we're *remixing*, the
original project should always have an identifier, so we don't need to
fallback to "project" like we do for saving.

I did wonder whether we should only do this if the remixed project
identifier (i.e. `action.payload.project.identifier`) is different from
the original project identifier (i.e. `state.project.identifier`).
However, I'm fairly confident this will always be true if we are
remixing.

Note that this came up when I was testing the changes in #992.

[1]: https://github.com/RaspberryPiFoundation/editor-ui/blob/70ea01cefa38aa60d1d6d9fb4bc11188b6009049/src/redux/EditorSlice.js#L373

Fixes #1001.