Fix project remix behaviour to match editor-ui
Previously, if a logged-in user had a remixed version of the project,
the editor web component would load it in preference to the original
version. The new `load_remix_disabled` attribute on the editor web
component [1] allows us to disable this behaviour in `editor-standalone`
while leaving it unchanged in the `projects-ui` app.

This moves us closer to being able to deploy `editor-standalone` in
place of `editor-ui` [2].

I did try to write an e2e spec to test this behaviour (see #85).
However, I ran into a bunch of problems the last of which was that while
`editor-standalone` was changed to use test-editor-api.raspberrypi.org
instead of staging-editor-api.raspberrypi.org, the web component being
served from staging-editor-static.raspberrypi.org continued to use
staging-editor-api.raspberrypi.org.

So in the end, I've resorted to writing an e2e spec for the web
component behaviour in RaspberryPiFoundation/editor-ui#1020.

[1]: https://github.com/RaspberryPiFoundation/editor-ui/pull/992
[2]: https://github.com/RaspberryPiFoundation/editor-standalone/issues/16