Fix warning re fill-available vs stretch
We were seeing multiple warnings when the app started up and therefore
also when running the Cypress specs in CI. The warnings were like this:

    Replace fill-available to stretch, because spec had been changed

The change from using "fill-available" to "stretch" in various CSS
dimension-related properties seems to have been a change to the CSS spec
a long time ago and I'm _fairly_ confident all browsers have supported
it for some time. Useful references:

* https://developer.mozilla.org/en-US/docs/Web/CSS/block-size
* https://developer.mozilla.org/en-US/docs/Web/CSS/width#browser_compatibility
* https://caniuse.com/?search=width%3A%20fill-available
* https://caniuse.com/?search=width%3A%20stretch

I've just done a global search & replace "fill-available" -> "stretch"
in this commit and this seems to fix the warnings and it doesn't seem to
have broken anything.