Rename ProjectName is/setEditable -> is/setEditing
This state tracks whether the user is in the process of editing the
project name or not and so I think the new name better reflects it's
purpose.

My actual motivation is that I want to more clearly distinguish
`isEditable` & `setEditable` from the `projectNameEditable` property
which I'm planning to rename to just `editable` in a subsequent commit.