Add project_name_editable attribute to web component (#1009)
Addresses #986 which is part of
RaspberryPiFoundation/editor-standalone#16.

This will allow us to make the project name editable in
`editor-standalone` (see RaspberryPiFoundation/editor-standalone#125)
while leaving other projects using the web component (e.g.
`projects-ui`) and the integrated version in `editor-ui` unchanged.

Note that this new attribute only affects the `ProjectName` component in
the `ProjectBar` and not the one in the sidebar, i.e. in
`ProjectsPanel`.