Remove unused `/embedded/projects/:identifier` route (#1013)
As far as we can tell, this route is [no longer used anywhere][1].

This was originally added in Mar 2022 in #72 for the "ESA demo". I
suspect we now achieve a similar effect by using the web component with
its `embedded` attribute [2] set to `true`, rather than using an `iframe`
with its `src` set to this route.

I found it confusing to see both the `/:locale/embed/viewer/:identifier`
and this route in the `AppRoutes` component. Removing this one makes the
code less confusing.

I've also removed the now redundant `embedded` property on the
`ProjectComponentLoader` component and the similarly redundant call to
`useEmbeddedMode` which is a no-op when `embedded` is `false`.

[1]: https://github.com/search?q=org%3Araspberrypilearning+embedded%2Fprojects&type=code
[2]: https://github.com/RaspberryPiFoundation/editor-ui/blob/340ab1192fb3358ad4614034208603948f8b95d6/src/web-component.js#L50