Add assets_identifier attribute to web component
Based on a curated set of changes from #901.

The new `assets_identifier` attribute is intended to be used in
conjunction with the `code` attribute so that the supplied source code
can reference images included in the project. Setting the attribute
results in a call to the `Api::Projects::ImagesController#show`
action [1] in `editor-api` which renders the
`app/views/api/projects/images.json.jbuilder` template [2] which makes
just the images for the project available; not the source files.

The motivation for this change is for the Block-to-Text app [3] which
uses the new attribute in its exercise to load the images for the
"bridge-prototype" project [4,5].

I've also added unit tests to test the new behaviour. I plan to add an
e2e spec in a subsequent commit which will test all the new attributes
on the web component.

[1]: https://github.com/RaspberryPiFoundation/editor-api/blob/96f435fbd9c2b1fb35821dd98e30a97056c6501a/app/controllers/api/projects/images_controller.rb#L8-L12
[2]: https://github.com/RaspberryPiFoundation/editor-api/blob/96f435fbd9c2b1fb35821dd98e30a97056c6501a/app/views/api/projects/images.json.jbuilder
[3]: https://github.com/RaspberryPiFoundation/block-to-text-alpha/
[4]: https://github.com/RaspberryPiFoundation/block-to-text-alpha/blob/d16a8ecc1d3da202363fb16db31f4bf93f46aa04/src/components/Editor.js#L15
[5]: https://github.com/raspberrypilearning/bridge-prototype