Enhance code attribute on web component
So that it overrides the content of the main component in the loaded
project. This means that for example, you can set the
`assets_identifier` to bring in some images from a real project, but
override the main component source code in `main.py` or `index.html`.

The motivation for this change is to support the Block-to-Text app [1].
Its exercise uses the `code` attribute to load the relevant code for
each step and modify it as the user makes changes [2].

This is based on a curated set of changes from #901. I've also added a
bunch of unit tests which hopefully explain a bit about the detail of
the behaviour. In doing that I noticed what I think is a small bug when
the real project has other non-default components (i.e. files other than
`main.py` or `index.html`), but I plan to fix that in a subsequent
commit.

I plan to add an e2e spec in a subsequent commit which will test all the
new attributes on the web component.

[1]: https://github.com/RaspberryPiFoundation/block-to-text-alpha/
[2]: https://github.com/RaspberryPiFoundation/block-to-text-alpha/blob/d16a8ecc1d3da202363fb16db31f4bf93f46aa04/src/components/Editor.js#L17