WIP: Expose embedded attribute on web component
TODO: Explain what web component `embedded` attribute does

Based on a curated set of changes from #901.

Note that in the original PR, an unnecessary call to
`dispatch(setEmbedded(embedded))` was added in `WebComponentLoader`.
This was unnecessary, because there was already a call to the
`useEmbeddedMode` hook which does the same thing, so I have removed it
from this commit.