Add e2e spec for wc in Block-to-Text mode
This adds a new e2e spec which loads the web component with its
attributes set to mimic the usage by Block-to-Text alpha app [1,2].

* The `assets_identifier` attribute is set to load the assets (but not
  the code) from the "encoded-art-starter" project which includes a
  `moon.png` image.
* The `host_styles` attribute is set partly because it is used by the
  Block-to-Text app, but also because otherwise the output-only editor
  has zero height in the Electron browser used by default by Cypress.
* The `code` attribute is set to load a simple P5 sketch which draws the
  word "PASS" in green and draws the `moon.png` image on the canvas.
* The `output_only` attribute is set to `true` so that only the output
  pane of the editor is displayed; not the code editor, the run bar,
  etc.
* The `output_panels` attribute is set to `["visual"]` so that only the
  visual output canvas is displayed; not the text output.
* A "Run code" button has been added to `src/web-component.html` to
  mimic a similar button in the Block-to-Text app which has an `onclick`
  handler to call the `rerunCode` method on the web component.

Note that the "embedded-viewer" CSS class that we're asserting against
is added by the `WebComponentProject` component around the `Output`
component; not by the `EmbeddedViewer` component which is currently only
used via the `/:locale/embed/viewer/:identifier` route which is used in
an iframe by e.g. the Projects app.

By asserting the results contain '{"errorDetails":{}}' we can have
_some_ confidence that the P5 sketch code has run successfully. However,
if e.g. the image is not available, the test will still pass despite an
uncaught exception being thrown. It might be worth investigating whether
Cypress can do visual regression testing to assert the canvas looks how
it is supposed to, but I'm going to say that's out-of-scope for the
moment.

[1]: https://block-to-text-alpha.pages.dev/
[2]: https://github.com/RaspberryPiFoundation/block-to-text-alpha/